+++ CHANGELOG.md
@@ -6,7 +6,10 @@ Notable changes will be documented in this file. The project adheres to [Semanti
 Unreleased
 ----------
 
+* Defined namespace in `ext_localconf` and `ext_tables`
 * Added `AbstractConditionViewHelper`
+* Updated code comments
+* Cleaned up
 
 2.0.3
 -----